azrp: first version of text shader (slow!), based on topti-cg

+++ azur/src/gint/shaders/rect.c
@@ -1,6 +1,6 @@
 #include <azur/gint/render.h>
 
-uint AZRP_SHADER_RECT = -1;
+uint8_t AZRP_SHADER_RECT = -1;
 
 static void configure(void)
 {
